+++ packages/ui-components/web/src/index.ts
@@ -48,4 +48,5 @@ export { default as FocusManager } from './Components/FocusManager';
 export { default as SettingsMobileHeader } from './Components/SettingsMobileHeader';
 export { default as SettingsSeparator } from './Components/SettingsSeparator';
 export { default as Label } from './Components/Label';
+export { default as SettingsPanelHeader } from './Components/SettingsPanelHeader';
 // export { getSelectionStyles } from './Utils/inlineStyleUtils';